refactor(tracking): move adapter to rock/sdk/job/tracking/ subpackage
- Move rock/sdk/job/adapter.py → rock/sdk/job/tracking/adapter.py
- Create rock/sdk/job/tracking/__init__.py package
- Fix docstring path: rock/sdk/tracking → rock/sdk/job/tracking
- Resolve adapters in Job.__init__ instead of on every report call
- Fix obj.__module__ check to use startswith (align with CommandLoader)
- Update all import paths in api.py, tests, and internal source
